Support incomplete entry
Allows a user to enter the rule if they did not specify in the original
command.
